Fix newline handling in template to prevent empty lines
The template for the xinetd.conf configuration file contains checks to
only generate entries for defined parameters. Unfortunately every end
tag is missing the hyphen notation to suppress a newline. This causes
the generated file to contain an empty line for every undefined
parameter. This patch updates the template tag to prevent unnecessary
empty lines in the output.