Make server parameter optional when redirect is set
According to the documentation, the server parameter is optional if a
redirection is configured. This patch makes the server parameter
optional. It includes a check to make sure at least one of the
parameters is defined. The manpage xinetd.conf(5) defines the precedence
rule if both parameters are defined, so that is not an error.