Allow service definition without explicit group
The group setting may be omitted in a service configuration to use the
users primary group. This config is currently not possible because
setting the group parameter to 'undef' will use the os specific default
group instead. This patch adds an additional parameter
'use_default_group' to control this behaviour. Using the default 'true'
keeps the old behaviour. Setting the parameter to 'false' and leaving
the group unset will omit the group setting from the service config
file.